first commit of somewhat-stable code (see comment below)
The previous development was very hectic and the code was not living in a standard
version control repository. Now the changes are starting to be more meaningful
(read: half of the files does not change within few minutes) and it makes sense to
put it into a Git repository to track the changes. It still does contain a lot of
development-related stuff, but most of the code should be "final".