URL reversing - closes #2787
Some notes about this code - first one - the speed of this implementation: it takes
approximately 200 ms to reverse 1000 routes in subapplication on Turris router. This
is negligible, only page it can be noticed on is the "about:config". Minor issue is that
it raises/catches same error when the name of the route can not be found and when
some URL arguments are missing (both cases raise RouteBuildError in build method).