remember last step in wizard (basic implementation) - refs #2839
Notes: code is not finished, some parts are missing (see TODO) and I didn't have time
to test it. Next allowed step is saved to session when appropriate, and to Uci config
when something is being already saved - this means that progress through time settings
and updater is saved only in session. But it's not a big deal, IMHO.