Font Awesome removed
It was thought that FA would be used more in the Foris, however the only place
it was used was the "about:config" page and only four icons of the whole set.
I've removed it and replaced icons with textual representation. If some font
icons are needed in future, we can make a custom web font to save space.