better config filter factory + taking advantage of it (see below)
In the past, there wasn't any large difference in requesting Get with filters
or without them, because of poor implementation of filters in Netconf.
However, since filtering is now smarter, requesting filtered data yields
a great deal of performance increase. This required a small change in config
filter factory, so it can create filter for more than one config.